Extract signing adapter factory
